- Merged files and added required imports.
